+++ Source/Math/Objects/EulerAngles.cpp
@@ -166,12 +166,12 @@ namespace Silent::Math
 
     EulerAngles EulerAngles::operator*(float scalar) const
     {
-        return EulerAngles((short)round((float)x * scalar), (short)round((float)y * scalar), (short)round((float)z * scalar));
+        return EulerAngles((short)roundf((float)x * scalar), (short)roundf((float)y * scalar), (short)roundf((float)z * scalar));
     }
 
     EulerAngles EulerAngles::operator/(float scalar) const
     {
-        return EulerAngles((short)round((float)x / scalar), (short)round((float)y / scalar), (short)round((float)z / scalar));
+        return EulerAngles((short)roundf((float)x / scalar), (short)roundf((float)y / scalar), (short)roundf((float)z / scalar));
     }
 
     short EulerAngles::InterpConstant(short from, short to, short angularVel)
@@ -193,7 +193,7 @@ namespace Silent::Math
         }
 
         short delta = to - from;
-        return (short)round(from + (delta * std::clamp(alpha, 0.0f, 1.0f)));
+        return (short)roundf(from + (delta * std::clamp(alpha, 0.0f, 1.0f)));
     }
 
     bool EulerAngles::Compare(short angle0, short angle1, short epsilon)
